refactor(orchestrator): extract recall search pipeline into orchestration/recall-search-pipeline.ts (seam 19, #1526)
Behavior-preserving move of the 10 QMD search post-processing methods
(query-aware prefilter, artifact top-up fetch, embedding/archive/cold
fallback pipelines, recall-safety filtering, memory-map loading, result
boosting) into RecallSearchPipelineCoordinator. 31-member deps bundle
wired with live accessors + late-binding arrows (same rule as seam 18);
orchestrator keeps thin delegating methods so private-API callers and
instance-level test stubs keep working. tokenizeRecallQuery exported
for the coordinator (seam-16/18 precedent).

orchestrator.ts: 9,489 -> 8,265 LOC.
